Save models account

+++ database/migrations/2021_03_12_201109_add_charges_to_users_table.php
@@ -14,7 +14,7 @@ class AddChargesToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->decimal('cherges')->default(0);
+            $table->integer('cherges')->default(0);
         });
     }
